scripts/requirements.txt
python-docx>=0.8.11
mistune>=3.0.0
lxml>=4.9.0
pyyaml>=6.0
